pyproject.toml
[project]
name = "recipe-creator"
version = "0.1.0"
description = "Recipe Creator Agent - AI-powered culinary assistant for personalized recipe generation, meal planning, and cooking guidance"
authors = [{ name = "Paras Chamoli", email = "paras25619@gmail.com" }]
readme = "README.md"
keywords = ['python', 'ai', 'agent', 'recipe', 'cooking', 'culinary', 'agno', 'bindu']
requires-python = ">=3.12,<3.13"
dependencies = [
    "bindu==2026.1.12",
    "agno>=2.2.0",
    "fastmcp>=2.11.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "openai>=2.11.0",
    "sqlalchemy>=2.0.45",
    "mem0ai>=1.0.1",
    "exa-py>=1.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Paraschamoli/recipe-creator"
Repository = "https://github.com/Paraschamoli/recipe-creator"
Documentation = "https://github.com/Paraschamoli/recipe-creator#readme"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["recipe_creator"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "YTT",
    "S",
    "B",
    "A",
    "C4",
    "T10",
    "SIM",
    "I",
    "C90",
    "E", "W",
    "F",
    "PGH",
    "UP",
    "RUF",
    "TRY",
]
ignore = [
    "E501",
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["recipe_creator"]
